src/requirements.txt
astroid==1.6.3
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.4
django-admin-sortable2==0.6.19
django-allauth==0.35.0
django-cache-machine==1.0.0
django-debug-toolbar==1.9.1
django-model-utils==3.1.2
django-rest-auth==0.9.3
djangorestframework==3.8.2
django-avatar==4.1.0
gunicorn==19.7.1
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.0.7
psycopg2==2.7.4
python-memcached==1.59
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
rest-framework-cache==0.1
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
whitenoise==3.3.1
wrapt==1.10.11
